Functions working.
Changed function calls to copy frames (which really
means copy the locals array so each call has it's
own set of local variables).
Tested some functions which accept multiple parameters,
return local variables, print globals, and so on.
A decent amount of clean-up and thought must go into
managing the function frames and local variables arrays
and so on.