Removed all unnecessary copies and destroys.
Now objects are only copied with a instructions that require it.
Object refcounts are only incremented when "stored" in a symbol,
and decremented when overwritten in a symbol by another object.

These changes have the effect of "leaking" memory for all objects
that allocate memory outside of the garbage collector (e.g. strings
allocate a `char *`, function frames allocate `Frame *`.