added "labels as values" utilization to interpreter.
Other bytecode interpreted languages like Python and Ruby
use GCC's "labels as values" extension (computed gotos).
It can improve the speed of the interpreter by up to 20%.
I describe the changes in interpet.c and provide links to both
Python's notes and notes from Eli Bendersky on the subject.

I am currently using macros to allow the code to compile as
a switch/case loop with compilers that don't support the
computed goto.