added static singleton LuciNilObj.
Changed many functions to return LuciNilObj rather than NULL.
This allows NULL to be reserved for errors, and 'nil' to be
used as an actual NULL type.

I'm not sure whether methods that a type doesn't implement should
return NULL and die, or return nil (e.g. list - list or map * map)...