large commit for garbage collection.
this commit is PROBABLY going to be a throwaway.
I did all that work on getting a SIMPLE stop-the-world
mark-and-sweep garbage collector working, only to realize
how many C-functions allocate containers and append objects
to them. This breaks if garbage collection runs while the
container's children are being allocated... ARGH.